Move SWV-ZFE tests to dedicated test file
